NHistory: refactor into separate file
